requirements.txt
bcrypt==3.2.0
cffi==1.14.3
colorama==0.4.4
cryptography==3.2.1
dataclasses==0.7
future==0.18.2
importlib-resources==3.3.0
mypy-extensions==0.4.3
netmiko==3.3.2
nornir==3.0.0
nornir-netmiko==0.1.1
nornir-utils==0.1.1
ntc-templates==1.6.0
paramiko==2.7.2
pkg-resources==0.0.0
pycparser==2.20
PyNaCl==1.4.0
pyserial==3.4
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
scp==0.13.3
six==1.15.0
tenacity==6.2.0
textfsm==1.1.0
typing-extensions==3.7.4.3
zipp==3.4.0
